+++ trunk/src/df1b2-separable/df1b2min2.cpp
@@ -41,7 +41,7 @@ void ad_read_pass2_minuscv(void);
 #endif
    //int nvar=df1b2variable::nvar;
 
-   int total_bytes=2*sizeof(df1b2_header);
+   size_t total_bytes=2*sizeof(df1b2_header);
 // string identifier debug stuff
 #if defined(SAFE_ALL)
   char ids[]="GU";
@@ -150,7 +150,7 @@ void read_pass2_2_minuscv(void)
   int nvar=df1b2variable::nvar;
   test_smartlist & list=f1b2gradlist->list;
 
-  int total_bytes=2*sizeof(df1b2_header);
+  size_t total_bytes=2*sizeof(df1b2_header);
 // string identifier debug stuff
 #if defined(SAFE_ALL)
   char ids[]="BY";
